+++ src/output/p_av1.cpp
@@ -55,8 +55,7 @@ av1_video_packetizer_c::set_header_parameters() {
   }
 
   if (!m_hvideo_display_width || !m_hvideo_display_height) {
-    set_video_display_width(dimensions.first);
-    set_video_display_height(dimensions.second);
+    set_video_display_dimensions(dimensions.first, dimensions.second, generic_packetizer_c::ddu_pixels, OPTION_SOURCE_BITSTREAM);
     need_to_rerender = true;
   }