+++ NEWS.md
@@ -38,6 +38,9 @@
 * mkvmerge: JSON identification: cropping parameters are reported again even
   if some of the four values are not present in the file or set to 0. Fixes
   #3534.
+* mkvmerge: AV1 packetizer: the packetizer did not honor the aspect ratio
+  (factor) given on the command-line via the `--aspect-ratio` or
+  `--aspect-ratio-factor` options.
 
 
 # Version 76.0 "Celebration" 2023-04-30